requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
drf-spectacular>=0.27,<1.0
django-ratelimit>=4.1,<5.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
uvicorn[standard]>=0.25,<1.0
redis>=5.0,<6.0
whitenoise>=6.6,<7.0
python-dotenv>=1.0,<2.0

pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-asyncio>=0.23,<1.0
factory-boy>=3.3,<4.0

ruff>=0.2,<1.0
bandit>=1.7,<2.0
pip-audit>=2.7,<3.0
pre-commit>=3.6,<4.0
